requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0  # FREE Gemini API!
langdetect
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.0.0
indic_transliteration